+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](
 - Currency name and description display and editing.
 - Show proposed venues on the venues page.
 - Render proposed venue content.
+- Allow additional support to a proposal.
 
 ## 0.0.4 2024-09-28